+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   `Screen::visual_bell_count`, and `Screen::errors` have been removed in favor
   of the new callback api described above.
 * `Cell` no longer implements `Default`.
+* `Screen` no longer implements `vte::Perform`.
 
 ### Changed